pyproject.toml
[project]
name = "make-llmstxt"
version = "0.1.0"
description = "Generate llms.txt and llms-full.txt files for any website using MCP scraper and LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
]

[project.scripts]
make-llmstxt = "make_llmstxt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
